refactor: transition to float32 precision across the codebase
- Updated the project to operate in float32 precision end to end, enhancing performance and memory efficiency.
- Adjusted environment settings to disable JAX's 64-bit mode by default.
- Refined type annotations and data handling to utilize EventArray consistently throughout the codebase.
- Improved documentation to reflect changes in precision handling and the rationale behind the transition from float64 to float32.
- Ensured compatibility across various modules by updating dtype specifications and removing legacy float64 references.

+++ tests/test_ibu.py
@@ -16,7 +16,7 @@ if TYPE_CHECKING:
 def _split(z: list[list[float]], x: list[list[float]], y: list[int]) -> ArrayDataset:
     return ArrayDataset(
         ZXY(
-            Events(np.asarray(z, dtype=np.double), np.asarray(x, dtype=np.double)),
+            Events(np.asarray(z, dtype=np.single), np.asarray(x, dtype=np.single)),
             np.asarray(y, dtype=np.ubyte),
         ),
         batch_size=8,
@@ -100,8 +100,8 @@ def test_parse_config_rejects_non_string_jet_variable() -> None:
 
 
 def test_assign_bins_saturates_underflow_and_overflow() -> None:
-    edges = np.array([0.0, 1.0, 2.0], dtype=np.double)
-    values = np.array([-3.0, 0.0, 0.4, 1.0, 2.0, 8.0], dtype=np.double)
+    edges = np.array([0.0, 1.0, 2.0], dtype=np.single)
+    values = np.array([-3.0, 0.0, 0.4, 1.0, 2.0, 8.0], dtype=np.single)
 
     indices = ibu._assign_bins(values, edges)
 
@@ -110,8 +110,8 @@ def test_assign_bins_saturates_underflow_and_overflow() -> None:
 
 
 def test_saturating_counts_conserve_observed_population() -> None:
-    edges = np.array([0.0, 1.0, 2.0], dtype=np.double)
-    observed = np.array([-2.0, 0.5, 4.0], dtype=np.double)
+    edges = np.array([0.0, 1.0, 2.0], dtype=np.single)
+    observed = np.array([-2.0, 0.5, 4.0], dtype=np.single)
 
     counts = ibu._bin_counts(ibu._assign_bins(observed, edges), 2)
 
@@ -156,8 +156,8 @@ def test_prepare_data_rejects_empty_test_data_population() -> None:
 
 def test_unfolded_to_bin_weights_uses_explicit_zero_prior_semantics() -> None:
     weights = ibu._unfolded_to_bin_weights(
-        np.array([4.0, 0.0], dtype=np.double),
-        np.array([2.0, 0.0], dtype=np.double),
+        np.array([4.0, 0.0], dtype=np.single),
+        np.array([2.0, 0.0], dtype=np.single),
     )
 
     np.testing.assert_array_equal(weights, [2.0, 0.0])
@@ -166,8 +166,8 @@ def test_unfolded_to_bin_weights_uses_explicit_zero_prior_semantics() -> None:
 def test_unfolded_to_bin_weights_rejects_mass_without_prior() -> None:
     with pytest.raises(ValueError, match="zero-prior"):
         ibu._unfolded_to_bin_weights(
-            np.array([4.0, 1.0], dtype=np.double),
-            np.array([2.0, 0.0], dtype=np.double),
+            np.array([4.0, 1.0], dtype=np.single),
+            np.array([2.0, 0.0], dtype=np.single),
         )
 
 
@@ -181,14 +181,14 @@ def test_unfolded_to_bin_weights_rejects_mass_without_prior() -> None:
     ],
 )
 def test_normalize_weights_rejects_invalid_vectors(
-    weights: NDArray[np.double], message: str
+    weights: NDArray[np.single], message: str
 ) -> None:
     with pytest.raises(ValueError, match=message):
         ibu._normalize_weights(weights)
 
 
 def test_normalize_weights_returns_mean_one() -> None:
-    normalized = ibu._normalize_weights(np.array([1.0, 2.0, 3.0], dtype=np.double))
+    normalized = ibu._normalize_weights(np.array([1.0, 2.0, 3.0], dtype=np.single))
 
     assert normalized.mean() == pytest.approx(1.0)
     assert np.all(normalized >= 0)
@@ -198,14 +198,14 @@ def testunfold_variable_reports_insufficient_bins_as_skip(monkeypatch) -> None:
     monkeypatch.setattr(
         ibu,
         "_purity_bins",
-        lambda *_args, **_kwargs: np.array([0.0, 1.0], dtype=np.double),
+        lambda *_args, **_kwargs: np.array([0.0, 1.0], dtype=np.single),
     )
 
     result = ibu.unfold_variable(
         variable_name="dim_0",
-        mc_gen=np.array([0.2, 0.8], dtype=np.double),
-        mc_sim=np.array([0.3, 0.7], dtype=np.double),
-        observed=np.array([-1.0, 2.0], dtype=np.double),
+        mc_gen=np.array([0.2, 0.8], dtype=np.single),
+        mc_sim=np.array([0.3, 0.7], dtype=np.single),
+        observed=np.array([-1.0, 2.0], dtype=np.single),
         n_iterations=2,
         purity_threshold=ibu.DEFAULT_PURITY_THRESHOLD,
     )
@@ -214,7 +214,7 @@ def testunfold_variable_reports_insufficient_bins_as_skip(monkeypatch) -> None:
     assert result.outcome.skip_reason == "fewer than two purity bins"
     assert result.reweighting is None
     np.testing.assert_array_equal(
-        result.weights_for(np.array([0.4, 0.6], dtype=np.double)), [1.0, 1.0]
+        result.weights_for(np.array([0.4, 0.6], dtype=np.single)), [1.0, 1.0]
     )
 
 
@@ -222,18 +222,18 @@ def testunfold_variable_returns_safe_mean_one_weights(monkeypatch) -> None:
     monkeypatch.setattr(
         ibu,
         "_purity_bins",
-        lambda *_args, **_kwargs: np.array([0.0, 1.0, 2.0], dtype=np.double),
+        lambda *_args, **_kwargs: np.array([0.0, 1.0, 2.0], dtype=np.single),
     )
 
     result = ibu.unfold_variable(
         variable_name="dim_0",
-        mc_gen=np.array([0.2, 0.8, 1.2, 1.8], dtype=np.double),
-        mc_sim=np.array([-1.0, 0.7, 1.3, 3.0], dtype=np.double),
-        observed=np.array([-2.0, 0.4, 1.4, 4.0], dtype=np.double),
+        mc_gen=np.array([0.2, 0.8, 1.2, 1.8], dtype=np.single),
+        mc_sim=np.array([-1.0, 0.7, 1.3, 3.0], dtype=np.single),
+        observed=np.array([-2.0, 0.4, 1.4, 4.0], dtype=np.single),
         n_iterations=2,
         purity_threshold=ibu.DEFAULT_PURITY_THRESHOLD,
     )
-    weights = result.weights_for(np.array([0.2, 1.8], dtype=np.double))
+    weights = result.weights_for(np.array([0.2, 1.8], dtype=np.single))
 
     assert result.outcome.status == "completed"
     assert result.outcome.n_bins == 2
@@ -252,8 +252,8 @@ def test_unfolds_in_single_precision_end_to_end() -> None:
     rng = np.random.default_rng(0)
     n = 20_000
     mc_gen = rng.normal(size=n).astype(np.single)
-    mc_sim = (mc_gen + 0.3 * rng.normal(size=n)).astype(np.single)
-    observed = (0.5 + rng.normal(size=n)).astype(np.single)
+    mc_sim = (mc_gen + 0.3 * rng.normal(size=n).astype(np.single)).astype(np.single)
+    observed = (0.5 + rng.normal(size=n).astype(np.single)).astype(np.single)
 
     result = ibu.unfold_variable(
         variable_name="dim_0",
@@ -271,7 +271,7 @@ def test_unfolds_in_single_precision_end_to_end() -> None:
 
     weights = result.weights_for(mc_gen)
     assert weights.dtype == np.single
-    assert weights.mean(dtype=np.double) == pytest.approx(1.0, rel=1e-5)
+    assert weights.mean(dtype=np.single) == pytest.approx(1.0, rel=1e-5)
 
 
 def test_unfolding_applies_to_a_sample_it_was_not_fit_on(monkeypatch) -> None:
@@ -279,14 +279,14 @@ def test_unfolding_applies_to_a_sample_it_was_not_fit_on(monkeypatch) -> None:
     monkeypatch.setattr(
         ibu,
         "_purity_bins",
-        lambda *_args, **_kwargs: np.array([0.0, 1.0, 2.0], dtype=np.double),
+        lambda *_args, **_kwargs: np.array([0.0, 1.0, 2.0], dtype=np.single),
     )
 
     result = ibu.unfold_variable(
         variable_name="dim_0",
-        mc_gen=np.array([0.2, 0.8, 1.2, 1.8], dtype=np.double),
-        mc_sim=np.array([-1.0, 0.7, 1.3, 3.0], dtype=np.double),
-        observed=np.array([-2.0, 0.4, 1.4, 4.0], dtype=np.double),
+        mc_gen=np.array([0.2, 0.8, 1.2, 1.8], dtype=np.single),
+        mc_sim=np.array([-1.0, 0.7, 1.3, 3.0], dtype=np.single),
+        observed=np.array([-2.0, 0.4, 1.4, 4.0], dtype=np.single),
         n_iterations=2,
         purity_threshold=ibu.DEFAULT_PURITY_THRESHOLD,
     )
@@ -295,7 +295,7 @@ def test_unfolding_applies_to_a_sample_it_was_not_fit_on(monkeypatch) -> None:
     assert result.reweighting.bin_weights.size == 2
     # Any number of events, in any order, so long as they are the same variable.
     for size in (2, 7, 50):
-        gen = np.linspace(0.0, 2.0, size, dtype=np.double)
+        gen = np.linspace(0.0, 2.0, size, dtype=np.single)
         assert result.weights_for(gen).shape == (size,)
 
 
@@ -309,7 +309,7 @@ def testunfold_variable_reports_population_count_mismatch(
     monkeypatch.setattr(
         ibu,
         "_purity_bins",
-        lambda *_args, **_kwargs: np.array([0.0, 1.0, 2.0], dtype=np.double),
+        lambda *_args, **_kwargs: np.array([0.0, 1.0, 2.0], dtype=np.single),
     )
     real_bin_counts = ibu._bin_counts
     call_count = 0
@@ -327,9 +327,9 @@ def testunfold_variable_reports_population_count_mismatch(
     with pytest.raises(ValueError, match=rf"{population}.*3.*4"):
         ibu.unfold_variable(
             variable_name="dim_0",
-            mc_gen=np.array([0.2, 0.8, 1.2, 1.8], dtype=np.double),
-            mc_sim=np.array([0.3, 0.7, 1.3, 1.7], dtype=np.double),
-            observed=np.array([0.1, 0.9, 1.1, 1.9], dtype=np.double),
+            mc_gen=np.array([0.2, 0.8, 1.2, 1.8], dtype=np.single),
+            mc_sim=np.array([0.3, 0.7, 1.3, 1.7], dtype=np.single),
+            observed=np.array([0.1, 0.9, 1.1, 1.9], dtype=np.single),
             n_iterations=2,
             purity_threshold=ibu.DEFAULT_PURITY_THRESHOLD,
         )
@@ -337,9 +337,9 @@ def testunfold_variable_reports_population_count_mismatch(
 
 def test_evaluate_dimension_accepts_one_dimensional_arrays() -> None:
     record = shared.evaluate_dimension(
-        reference=np.array([0.0, 1.0, 2.0], dtype=np.double),
-        comparison=np.array([0.0, 1.5, 3.0], dtype=np.double),
-        weights=np.ones(3, dtype=np.double),
+        reference=np.array([0.0, 1.0, 2.0], dtype=np.single),
+        comparison=np.array([0.0, 1.5, 3.0], dtype=np.single),
+        weights=np.ones(3, dtype=np.single),
     )
 
     assert set(record) == {
@@ -361,7 +361,7 @@ def test_run_and_evaluate_returns_named_aligned_result(monkeypatch) -> None:
     monkeypatch.setattr(
         ibu,
         "_purity_bins",
-        lambda *_args, **_kwargs: np.array([0.0, 1.0], dtype=np.double),
+        lambda *_args, **_kwargs: np.array([0.0, 1.0], dtype=np.single),
     )
     config = shared.parse_run_config(_config(dim=1, gaussian_params={"dim": 1}))
 
@@ -388,7 +388,7 @@ def test_run_and_evaluate_returns_named_aligned_result(monkeypatch) -> None:
         (True, ibu.DEFAULT_PURITY_THRESHOLD, "n_iterations"),
         (1, np.double(np.nan), "purity_threshold"),
         (1, np.double(-0.1), "purity_threshold"),
-        (1, np.double(1.1), "purity_threshold"),
+        (1, np.single(1.1), "purity_threshold"),
     ],
 )
 def test_run_and_evaluate_validates_controls_before_loading_data(
